# Improve interop performance against CLR list types