--- a/Jint/Runtime/Interop/Reflection/FieldAccessor.cs
+++ b/Jint/Runtime/Interop/Reflection/FieldAccessor.cs
@@ -1,27 +1,26 @@
 using System.Reflection;
 
-namespace Jint.Runtime.Interop.Reflection
+namespace Jint.Runtime.Interop.Reflection;
+
+internal sealed class FieldAccessor : ReflectionAccessor
 {
-    internal sealed class FieldAccessor : ReflectionAccessor
+    private readonly FieldInfo _fieldInfo;
+
+    public FieldAccessor(FieldInfo fieldInfo, PropertyInfo? indexer = null)
+        : base(fieldInfo.FieldType, indexer)
     {
-        private readonly FieldInfo _fieldInfo;
+        _fieldInfo = fieldInfo;
+    }
 
-        public FieldAccessor(FieldInfo fieldInfo, string? memberName = null, PropertyInfo? indexer = null)
-            : base(fieldInfo.FieldType, memberName, indexer)
-        {
-            _fieldInfo = fieldInfo;
-        }
+    public override bool Writable => !_fieldInfo.Attributes.HasFlag(FieldAttributes.InitOnly);
 
-        public override bool Writable => !_fieldInfo.Attributes.HasFlag(FieldAttributes.InitOnly);
+    protected override object? DoGetValue(object target, string memberName)
+    {
+        return _fieldInfo.GetValue(target);
+    }
 
-        protected override object? DoGetValue(object target)
-        {
-            return _fieldInfo.GetValue(target);
-        }
-
-        protected override void DoSetValue(object target, object? value)
-        {
-            _fieldInfo.SetValue(target, value);
-        }
+    protected override void DoSetValue(object target, string memberName, object? value)
+    {
+        _fieldInfo.SetValue(target, value);
     }
 }