--- a/Jint/Runtime/Interop/Reflection/MethodAccessor.cs
+++ b/Jint/Runtime/Interop/Reflection/MethodAccessor.cs
@@ -1,30 +1,28 @@
 using Jint.Runtime.Descriptors;
 
-namespace Jint.Runtime.Interop.Reflection
+namespace Jint.Runtime.Interop.Reflection;
+
+internal sealed class MethodAccessor : ReflectionAccessor
 {
-    internal sealed class MethodAccessor : ReflectionAccessor
+    private readonly Type _targetType;
+    private readonly MethodDescriptor[] _methods;
+
+    public MethodAccessor(Type targetType, MethodDescriptor[] methods) : base(null!)
     {
-        private readonly Type _targetType;
-        private readonly string _name;
-        private readonly MethodDescriptor[] _methods;
+        _targetType = targetType;
+        _methods = methods;
+    }
 
-        public MethodAccessor(Type targetType, string name, MethodDescriptor[] methods)
-            : base(null!, name)
-        {
-            _targetType = targetType;
-            _name = name;
-            _methods = methods;
-        }
+    public override bool Writable => false;
 
-        public override bool Writable => false;
+    protected override object? DoGetValue(object target, string memberName) => null;
 
-        protected override object? DoGetValue(object target) => null;
+    protected override void DoSetValue(object target, string memberName, object? value)
+    {
+    }
 
-        protected override void DoSetValue(object target, object? value) { }
-
-        public override PropertyDescriptor CreatePropertyDescriptor(Engine engine, object target, bool enumerable = true)
-        {
-            return new(new MethodInfoFunction(engine, _targetType, target, _name, _methods), PropertyFlag.AllForbidden);
-        }
+    public override PropertyDescriptor CreatePropertyDescriptor(Engine engine, object target, string memberName, bool enumerable = true)
+    {
+        return new(new MethodInfoFunction(engine, _targetType, target, memberName, _methods), PropertyFlag.AllForbidden);
     }
 }