… Jint/Runtime/Interop/Reflection/IndexerAccessor.cs
… Jint/Runtime/Interop/Reflection/NestedTypeAccessor.cs
… Jint/Runtime/Interop/Reflection/PropertyAccessor.cs
… Jint/Runtime/Interop/Reflection/ReflectionAccessor.cs
… Jint/Runtime/Interop/TypeReference.cs
… Jint/Runtime/Interop/TypeResolver.cs