fixes #1772 


## Jint.Benchmark.ListInteropBenchmark

| **Diff**|Method|Toolchain|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Filter|Default|967.84 μs|12.909 μs|1935.49 KB|
| **New** |	| **Default** | **704.80 μs (-27%)** | **13.149 μs** | **973936 B (-51%)** |
| Old |Indexing|Default|305.84 μs|4.030 μs|113.96 KB|
| **New** |	| **Default** | **263.17 μs (-14%)** | **3.922 μs** | **92667 B (-21%)** |
| Old |HasProperty|Default|18.40 μs|0.246 μs|30.97 KB|
| **New** |	| **Default** | **11.17 μs (-39%)** | **0.090 μs** | **32 B (-100%)** |